Assignment_08/requirements.txt
pandas
sentence-transformers
faiss-cpu
transformers
torch 
kagglehub 